pyproject.toml
[project]
name = "qrt-challenge-2025"
version = "0.1.0"
description = "https://challengedata.ens.fr/participants/challenges/162/"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.11.12",
    "autoviz>=0.1.905",
    "deepsurv>=0.1.0",
    "featuretools[complete]>=1.31.0",
    "graphviz>=0.20.3",
    "ipykernel>=6.29.5",
    "lifelines>=0.30.0",
    "lightgbm>=4.5.0",
    "matplotlib>=3.9.4",
    "myvariant>=1.0.0",
    "numpy==1.26.4",
    "pandas>=2.2.3",
    "pip>=24.3.1",
    "pycox>=0.3.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "scikit-learn>=1.5.2",
    "scikit-survival>=0.23.1",
    "setuptools>=75.8.0",
    "shap>=0.46.0",
    "sklearn-pandas>=2.2.0",
    "sweetviz==2.2.1",
    "tensorflow>=2.18.0",
    "torch>=2.6.0",
    "torchtuples>=0.2.2",
]
